Refactor RunMode
